requirements.txt
alive-progress
numpy<2
geopandas
psutil
pyproj
scikit-image
requests
zenodo_search

gdal[numpy]

git+https://github.com/quantifyearth/iucn_modlib
git+https://github.com/quantifyearth/pyshark
git+https://github.com/quantifyearth/yirgacheffe@cc89b4d8a0e97c1a11b730cd688a58b680023336

pylint
mypy
pytest
